requirements/requirements.txt
-i https://pypi.org/simple
about-time==4.2.1; python_version >= '3.7' and python_version < '4'
alive-progress==3.1.5; python_version >= '3.7' and python_version < '4'
annotated-types==0.6.0; python_version >= '3.8'
attrs==23.2.0; python_version >= '3.7'
certifi==2023.11.17; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
filelock==3.13.1; python_version >= '3.8'
fsspec==2023.12.2; python_version >= '3.8'
gitdb==4.0.11; python_version >= '3.7'
gitpython==3.1.40; python_version >= '3.7'
grapheme==0.6.0
h5py==3.10.0; python_version >= '3.8'
huggingface-hub==0.20.2; python_full_version >= '3.8.0'
idna==3.6; python_version >= '3.5'
jinja2==3.1.2; python_version >= '3.7'
jsonlines==4.0.0; python_version >= '3.8'
markupsafe==2.1.3; python_version >= '3.7'
mpmath==1.3.0
networkx==3.2.1; python_version >= '3.9'
numpy==1.26.3; python_version >= '3.9'
packaging==23.2; python_version >= '3.7'
psutil==5.9.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pydantic==2.5.3; python_version >= '3.7'
pydantic-core==2.14.6; python_version >= '3.7'
pyyaml==6.0.1; python_version >= '3.6'
regex==2023.12.25; python_version >= '3.7'
requests==2.31.0; python_version >= '3.7'
safetensors==0.4.1; python_version >= '3.7'
-e .
smmap==5.0.1; python_version >= '3.7'
sympy==1.12; python_version >= '3.8'
tabulate==0.9.0; python_version >= '3.7'
tokenizers==0.15.0; python_version >= '3.7'
torch==2.1.2; python_full_version >= '3.8.0'
tqdm==4.66.1; python_version >= '3.7'
transformers==4.36.2; python_full_version >= '3.8.0'
types-tabulate==0.9.0.3
typing-extensions==4.9.0; python_version >= '3.8'
urllib3==2.1.0; python_version >= '3.8'
